+++ src/mafft/core/setdirection.c
@@ -8,11 +8,7 @@ static int show_R_ = 1;
 static int subalignment;
 static int subalignmentoffset;
 
-#ifdef ismodule
-void arguments_setdirection( int argc, char *argv[] )
-#else
-void arguments( int argc, char *argv[] )
-#endif
+static void arguments( int argc, char *argv[] )
 {
     int c;
 
@@ -81,11 +77,7 @@ int setdirection( int argc, char *argv[] )
 	char firstdir;
 	char *directions;
 
-#ifdef ismodule
-	arguments_setdirection( argc, argv );
-#else
 	arguments( argc, argv );
-#endif
 
 	reporterr( "subalignment = %d\n", subalignment );
 	reporterr( "subalignmentoffset = %d\n", subalignmentoffset );